examples/zspin-cloud-saas/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pyjwt==2.10.1
stripe==13.1.1
